[Next Gen] refactor(codegen): split _createSlots construction into create_slots
